Refactor Template HTML
