+++ .env_template
@@ -12,5 +12,5 @@ MAIL_PORT=
 MAIL_FROM=noreply@domaine
 MAIL_FROM_NAME=
 
-# WEB_URL doit correspondre au chemin du dossier racine (important surtout en local)
+# WEB_URL nom du virtualhost
 WEB_URL =
