+++ index.js
@@ -3,7 +3,6 @@ const sendBtn = document.querySelector("#send");
 const eMailDom = document.querySelector('#destEmail');
 const eMailListDom = document.querySelector('.email-list');
 const eMailAddDom = document.querySelector('.email-add');
-const weburl = document.getElementById('weburl').value;
 
 // Vérifie si l'email est correct
 const isEmailValid = (email) => {
@@ -14,7 +13,7 @@ const isEmailValid = (email) => {
 }
 
 // Défini l'URL et le formulaire
-const url = weburl + `/src/upload.php`;
+const url = `${location.origin}/src/upload.php`;
 const form = document.querySelector('form');
 
 function isEmptyFile() {
@@ -24,6 +23,21 @@ function isEmptyFile() {
         $('#modalEmptyFile1').modal('show');
     }
 }
+
+// Affiche un spinner
+function displaySpinner(){
+
+    setTimeout(function(){
+        document.getElementById("spin").hidden = false;
+    });
+
+    setTimeout(function(){
+        document.getElementById("spin").hidden = true;
+    }, 5000);
+
+}
+
+ 
 
 function updateFileName() {
     const input = document.getElementById('fichier');
@@ -122,6 +136,8 @@ function eMailCountDomUpdate() {
 form.addEventListener('submit', async (event) => {
     event.preventDefault();
 
+    displaySpinner();
+    isEmptyFile();
     // Déclaration des variables    
 
     const destEmail = [...document.querySelector('.email-list').childNodes].map((value) => {
@@ -155,8 +171,7 @@ form.addEventListener('submit', async (event) => {
     eMailListDom.replaceChildren();
     eMailCountDomUpdate();
 
-    displaySpinner();
-    isEmptyFile();
+
 })
 
 // Effet du nom du site
