+++ index.php
@@ -6,19 +6,11 @@ use Dotenv\Dotenv;
 
 $dotenv = Dotenv::createImmutable(__DIR__);
 $dotenv->load();
-
+$title = 'EasyUpload';
+include 'src/_header.php';
 ?>
 
-<!DOCTYPE html>
-<html lang="en">
 
-<head>
-	<meta charset="UTF-8">
-	<meta name="viewport" content="width=device-width, initial-scale=1.0">
-	<title><?= $_ENV['MAIL_FROM_NAME'] ?></title>
-	<link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-QWTKZyjpPEjISv5WaRU9OFeRpok6YctnYmDr5pNlyT2bRjXh0JMhjY6hW+ALEwIH" crossorigin="anonymous">
-	<link rel="stylesheet" href="style.css">
-</head>
 
 <body>
 	<div>
@@ -76,7 +68,7 @@ $dotenv->load();
 
 					<button class="email-add">🞦</button>
 					<!-- use attribute `multiple` ? -->
-					<input type="email" placeholder="example@example.com Enter" value='' class="form-control custom-input" id="destEmail" name="destEmail">
+					<input type="email" placeholder="example@example.com" value='' class="form-control custom-input" id="destEmail" name="destEmail">
 				</div>
 				<div class="mb-3">
 					<label for="sourceEmail" class="form-label">Votre Email</label>
@@ -89,12 +81,6 @@ $dotenv->load();
 			</form>
 		</div>
 	</div>
-	<script src="https://code.jquery.com/jquery-3.2.1.slim.min.js" integrity="sha384-KJ3o2DKtIkvYIK3UENzmM7KCkRr/rE9/Qpg6aAZGJwFDMVNA/GpGFF93hXpG5KkN" crossorigin="anonymous"></script>
-	<script src="https://cdn.jsdelivr.net/npm/popper.js@1.12.9/dist/umd/popper.min.js" integrity="sha384-ApNbgh9B+Y1QKtv3Rn7W3mgPxhU9K/ScQsAP7hUibX39j7fakFPskvXusvfa0b4Q" crossorigin="anonymous"></script>
-	<script src="https://cdn.jsdelivr.net/npm/bootstrap@4.0.0/dist/js/bootstrap.min.js" integrity="sha384-JZR6Spejh4U02d8jOt6vLEHfe/JQGiRRSQQxSfFWpi1MquVdAyjUar5+76PVCmYl" crossorigin="anonymous"></script>
-	<script src='index.js'></script>
-	<script src='script.js'></script>
-
-</body>
-
-</html>
+<?php
+include 'src/_footer.php';
+?>
