--- info.php
@@ -1,8 +1,0 @@
-<?php
-phpinfo();
-
-$var1 = [1, 71, 55, 32, 19];
-
-foreach ($var1 as $value) {
-    echo $value;
-}
