+++ src/logout.php
